chore(refactoring): small refactoring
