--- syncano/helpers/envs.js
@@ -1,7 +1,0 @@
-const envs = {
-  synqApiKey: '',
-  synqUrl: 'https://api.synq.fm/v1/',
-  syncanoUrl: 'https://api.syncano.rocks/v2/'
-}
-
-export default envs
