+++ syncano/synq/helpers/envs.js
@@ -1,0 +1,7 @@
+const envs = {
+  synqApiKey: CONFIG.SYNQ_API_KEY,
+  synqUrl: 'https://api.synq.fm/v1/',
+  syncanoUrl: `https://${META.request.HTTP_HOST}/v2/`
+}
+
+export default envs
