+++ syncano/synq/scripts/delete.js
@@ -1,7 +1,7 @@
-import db from '../../helpers/db'
-import respond from '../../helpers/respond'
 import fetch from 'node-fetch'
 import FormData from 'form-data'
+import db from '../helpers/db'
+import respond from '../helpers/respond'
 
 const videoObject = ARGS.POST
 const user = META.user || {}
